Refactor stats hooks
