+++ src/app/stats/components/HntInfo.tsx
@@ -7,7 +7,9 @@ import { MOBILE_MINT, amountAsNum, toNumber } from "@helium/spl-utils"
 import { format } from "date-fns"
 import { useAsync } from "react-async-hook"
 import Countdown from "react-countdown"
-import { useRegistrar, useSubDaoEpochInfo, useUnixTimestamp } from "../hooks"
+import { useRegistrar } from "../hooks/useRegistrar"
+import { useSubDaoEpochInfo } from "../hooks/useSubDaoEpochInfo"
+import { useUnixTimestamp } from "../hooks/useUnixTimestamp"
 import { ONE_DAY_MS, fetcher } from "../utils"
 import { CountdownRenderer, StatItem } from "./StatItem"
 import { StatsList } from "./StatsList"
